movies_admin/requirements/base.txt
asgiref
Django==3.2
django-debug-toolbar==2.2
django-extensions==3.1.3
django-filter==2.4.0
django-model-utils==4.0.0
djangorestframework==3.12.4
factory-boy==3.2.0
Faker==8.11.0
Markdown==3.3.4
psycopg2-binary==2.9.1
python-dateutil==2.8.2
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
text-unidecode==1.3
drf-yasg==1.20.0
python-decouple
